Update to try and remove fragment properly
The spiral fragment was being kept alive by the infinite animation,
hence why it wasn't being removed from the fragment container. The fade
transition in the fragment manager was one way to handle this, but
removing the fragment altogether was the initial aim.
Added a slight delay before starting a new experiment.
The bounce experiment will now limit itself a bit better when asked to
stop.